更新内容： 1.封装最新版本AsyncHttpClient网络请求,采用接口回调模式。（删除之前两个jar） 2.数据解析采用fastjson正反序列化。 3.更新Auth接口（getTokenId、login、logout）
说明：为了程序能够运行，这里所有关于修改API相关的类，都是重新创建，名字再其后面加1（...1）。

+++ app/src/main/java/com/pictureair/photopass/activity/CartActivity.java
@@ -33,7 +33,6 @@ import com.pictureair.photopass.widget.MyToast;
 import com.pictureair.photopass.widget.NoNetWorkOrNoCountView;
 import com.pictureair.photopass.widget.XListViewHeader;
 
-import org.apache.http.Header;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -43,6 +42,8 @@ import java.io.FileNotFoundException;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
+
+import cz.msebera.android.httpclient.Header;
 
 /**
  * 购物车页面
@@ -258,7 +259,7 @@ public class CartActivity extends BaseActivity implements OnClickListener {
 					}
 					
 					@Override
-					public void onProgress(int bytesWritten, int totalSize) {
+					public void onProgress(long bytesWritten, long totalSize) {
 						super.onProgress(bytesWritten, totalSize);
 						dialog.setProgress(bytesWritten, totalSize);
 					};
